Set up Vite for building and testing
