+++ package.json
@@ -1,17 +1,34 @@
 {
   "name": "@ivteplo/bottom-sheet",
+  "description": "Bottom sheet component, implemented in pure JavaScript",
+  "author": "Ivan Teplov <iteplov7@gmail.com>",
+  "license": "Apache-2.0",
   "version": "1.0.0",
-  "description": "Bottom sheet, implemented in pure HTML, CSS, and JavaScript",
-  "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+
+  "type": "module",
+  "files": ["build"],
+  "main": "./build/index.umd.cjs",
+  "module": "./build/index.js",
+  "exports": {
+    ".": {
+      "require": "./build/index.umd.cjs",
+      "import": "./build/index.js"
+    }
   },
+
+  "scripts": {
+    "dev": "vite --config test/vite.config.js",
+    "build": "vite build"
+  },
+
+  "devDependencies": {
+    "vite": "^4.3.0"
+  },
+
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ivteplo/bottom-sheet.git"
   },
-  "author": "Ivan Teplov <iteplov7@gmail.com>",
-  "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/ivteplo/bottom-sheet/issues"
   },
