+++ test/index.html
@@ -14,7 +14,7 @@
     <link rel="stylesheet" href="main.css">
 
     <!-- Script that makes the sheet work -->
-    <script src="main.js" defer></script>
+    <script type="module" src="main.js" defer></script>
   </head>
   <body class="font-system column fill items-center justify-center">
     <!-- Button to open a sheet -->
